requirements.ws.txt
torch
numpy==2.0.2
datasets==3.2.0
loguru==0.7.3
mosaicml-streaming==0.10.0
sentence-transformers==3.4.1
PyYAML==6.0.2
transformers==4.48.1
wandb==0.19.1
accelerate==1.3.0
deepspeed==0.16.3